Cleaned up Parser and moved all the create methods to the factory
