+++ golang/src/main/java/com/oracle/app/parser/Parser.java
@@ -60,8 +60,7 @@ public class Parser {
 				}
 			}
 		}
-		return allFunctions;
-		
+		return factory.getAllFunctions();
 	}
 	
 	private GoStatementNode recParse(String currNode) throws IOException {
@@ -91,38 +90,31 @@ public class Parser {
 	    		GoStatementNode par = recParse(nodeType);
 	    		if(par != null)
 	    			body.add(par);
-	    		
-	    		
 	    	}
 	    	else {
 	    		matchedTerm = attr.matcher(currentLine);
 	    		matchedTerm.find();
 	    		bindex = matchedTerm.start();
 	    		attrs.add(currentLine.substring(bindex));
-	    		
 	    	}
 		}
-		
 		return null;
 	}
 
-
 	public GoStatementNode getNodeType(String nodeType, ArrayList<String> attrs, ArrayList<GoStatementNode> body) throws IOException{
-		//System.out.println(nodeType + " " + body.size());
+
 		String name = searchAttr("Name: ", attrs);
 		String value = searchAttr("Value: ", attrs);
 		switch(nodeType) {
 			case "File":
 				return new GoBasicNode(nodeType, body.toArray(new GoStatementNode[body.size()]));
+				
 			case "Ident":
-				//Definitely will need changing but is going to be used for call expr for now
-				//without chained calls
-				//TEMPORARY PLS CHANGE
 				return new GoFunctionLiteralNode(language, name);
 				
 			case "Decl":
 				//Start a new lexical scope for decls
-				return decl(body);
+				return factory.createDecl(body);
 				
 			case "Spec":
 				System.out.println(nodeType);
@@ -133,11 +125,11 @@ public class Parser {
 				break;
 				
 			case "BasicLit":
-				return basicLit(value);
+				return factory.createBasicLit(value);
 				
 			case "FuncDecl":
 				//Start a new lexical scope
-				createFunction(name, body);
+				factory.createFunction(name, body);
 				break;
 				
 			case "Object":
@@ -148,22 +140,25 @@ public class Parser {
 				
 			case "BlockStmt":
 				//needs to return a block node
-				return createFunctionBlock(body);
+				return factory.createFunctionBlock(body);
 				
 			case "Stmt":
 				return new GoBasicNode(nodeType, body.toArray(new GoExpressionNode[body.size()]));
 				
 			case "ExprStmt":
 				return new GoBasicNode(nodeType, body.toArray(new GoExpressionNode[body.size()]));
+				
 			case "CallExpr":
 				//Create invoke node
-				return createInvoke(body);
+				return factory.createInvoke(body);
+				
 			case "SelectorExpr":
 				System.out.println(nodeType);
 				break;
 				
 			case "Expr":
-				return createExpr(body);
+				return factory.createExpr(body);
+				
 			case "GenDecl":
 				return new GoBasicNode(nodeType, body.toArray(new GoExpressionNode[body.size()]));
 				
@@ -178,7 +173,6 @@ public class Parser {
 				
 		}
 		return null;
-
 	}
 	
 	public String searchAttr(String attr, ArrayList<String> attrs) {
@@ -192,90 +186,11 @@ public class Parser {
 			name = name.split(": ", 2)[1];
 			name = name.substring(1, name.length()-1);
 		}
-		//System.out.println(name);
 		return name;
-	}
-	
-	public GoExpressionNode basicLit(String name) throws IOException{
-		return new GoStringNode(name);
-	}
-	
-	public GoExpressionNode createExpr(ArrayList<GoStatementNode> body) throws IOException{
-		return new GoExprNode(body.toArray(new GoExpressionNode[body.size()]));
-	}
-	
-	public GoInvokeNode createInvoke(ArrayList<GoStatementNode> body) throws IOException{
-		GoExpressionNode function = (GoExpressionNode) body.remove(0);
-		GoInvokeNode invokeNode = new GoInvokeNode(function, body.toArray(new GoExpressionNode[body.size()]));
-		return invokeNode;
-	}
-	
-	/*
-	 * Create a block of statements. Currently only specific to function blocks
-	 */
-	
-	public GoFunctionBodyNode createFunctionBlock(ArrayList<GoStatementNode> body) throws IOException{
-		
-		GoBlockNode blockNode = new GoBlockNode(body.toArray(new GoStatementNode[body.size()]));
-		GoFunctionBodyNode bodyNode = new GoFunctionBodyNode(blockNode);
-		return bodyNode;
-	} 
-	
-	private void flattenBlocks(Iterable<? extends GoStatementNode> bodyNodes, List<GoStatementNode> flatNodes){
-		for(GoStatementNode node : bodyNodes){
-			if(node instanceof GoBlockNode){
-				flattenBlocks(((GoBlockNode) node).getStatements(), flatNodes);
-			}
-			else{
-				flatNodes.add(node);
-			}
-		}
-	}
-	
-	/*
-	 * Creates a function node and adds it to the function hashmap
-	 * Needs to still add in paramters and return types/parameters
-	 * and lexical scope is needed too
-	 */
-	public void createFunction(String name, ArrayList<GoStatementNode> body) throws IOException{
-/*
-		String name = searchAttr("Name: ", attrs);
-		System.out.println("name in create function: " + name);
-		for(int i = 0; i < body.size(); i++) {
-			if(body.get(i) instanceof GoFunctionBodyNode) {
-				System.out.println("FOUND THE BODY NODE " + i);
-			}
-		}
-		*/
-		System.out.println("name in create function: " + name);
-		GoRootNode root = new GoRootNode(language,null,(GoFunctionBodyNode)body.get(2),null,"main");
-		allFunctions.put("main",root);
-	}
-	
-	/*
-	 * Currently only used by createFunction to get the function name
-	 * Working only off of the assumption of a HelloWorld main function
-	 * A BUNCH OF ASSUMPTIONS BUT EH..... jk should be fixed
-	 */
-	public String ident() throws IOException{
-		reader.readLine();
-		currentLine = reader.readLine();
-		return currentLine.split("\"")[1];
-	}
-	
-	public GoStatementNode[] genDecl() throws IOException{
-		return null;
-	}
-	
-	public GoDeclNode decl(ArrayList<GoStatementNode> body) throws IOException{
-		GoDeclNode node = new GoDeclNode(body.toArray(new GoStatementNode[body.size()]));
-		return node;
 	}
 	
 	public static Map<String, GoRootNode> parseGo(GoLanguage language, Source source) throws IOException{
 		Parser parser = new Parser(language, source);
 		return parser.beginParse();
 	}
-	
-
 }
